backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-dotenv==1.0.1
google-genai==1.7.0
pydantic==2.9.2
pydantic-settings==2.5.2
